extracted JSON to the separate bundle and feature
