+++ bundles/ru.arsysop.passage.lbc.base/src/ru/arsysop/passage/lbc/base/request/AdminRequestExecutor.java
@@ -29,7 +29,7 @@ import javax.servlet.http.HttpServletResponse;
 import ru.arsysop.passage.lbc.base.BaseComponent;
 import ru.arsysop.passage.lbc.server.ServerRequestAction;
 import ru.arsysop.passage.lbc.server.ServerRequestExecutor;
-import ru.arsysop.passage.lic.net.RequestParameters;
+import ru.arsysop.passage.lic.net.LicensingRequests;
 
 public class AdminRequestExecutor extends BaseComponent implements ServerRequestExecutor {
 
@@ -43,7 +43,7 @@ public class AdminRequestExecutor extends BaseComponent implements ServerRequest
 	public void executeRequest(HttpServletRequest request, HttpServletResponse response)
 			throws IOException, ServletException {
 
-		String actionId = request.getParameter(RequestParameters.ACTION);
+		String actionId = request.getParameter(LicensingRequests.ACTION);
 		ServerRequestAction requestAction = mapActionRequest.get(actionId);
 		if (requestAction != null) {
 			requestAction.execute(request, response);
@@ -54,7 +54,7 @@ public class AdminRequestExecutor extends BaseComponent implements ServerRequest
 
 	@Override
 	public boolean checkAccesstMode(HttpServletRequest baseRequest) {
-		String requestAccessMode = baseRequest.getParameter(RequestParameters.MODE);
+		String requestAccessMode = baseRequest.getParameter(LicensingRequests.MODE);
 		if (requestAccessMode != null && requestAccessMode.equals(accessModeId)) {
 			return true;
 		}
