+++ bundles/ru.arsysop.passage.lbc.container/src/ru/arsysop/passage/lbc/container/ServerRunnerImpl.java
@@ -32,7 +32,7 @@ import ru.arsysop.passage.lbc.server.ServerRequestAction;
 import ru.arsysop.passage.lbc.server.ServerRequestExecutor;
 import ru.arsysop.passage.lbc.server.ServerRequestHandler;
 import ru.arsysop.passage.lbc.server.ServerRunner;
-import ru.arsysop.passage.lic.net.RequestParameters;
+import ru.arsysop.passage.lic.net.LicensingRequests;
 
 public class ServerRunnerImpl implements ServerRunner {
 
@@ -74,7 +74,7 @@ public class ServerRunnerImpl implements ServerRunner {
 
 	public void bindServerRequestHandler(ServerRequestHandler serverRequestHandler, Map<String, String> context) {
 		logger.debug(String.format(REGISTERED, serverRequestHandler.getClass().getName()));
-		String requestHandlerId = context.get(RequestParameters.HANDLER);
+		String requestHandlerId = context.get(LicensingRequests.HANDLER);
 		if (requestHandlerId != null) {
 			requestHandlers.put(requestHandlerId, serverRequestHandler);
 			serverHandler.addServerRequestHandler(serverRequestHandler);
@@ -84,7 +84,7 @@ public class ServerRunnerImpl implements ServerRunner {
 	}
 
 	public void unbindServerRequestHandler(ServerRequestHandler serverRequestHandler, Map<String, String> context) {
-		String requestHandlerId = context.get(RequestParameters.HANDLER);
+		String requestHandlerId = context.get(LicensingRequests.HANDLER);
 		if (requestHandlerId != null) {
 			requestHandlers.remove(requestHandlerId, serverRequestHandler);
 			serverHandler.remServerRequestHandler(serverRequestHandler);
@@ -95,7 +95,7 @@ public class ServerRunnerImpl implements ServerRunner {
 
 	public void bindServerRequestExecutor(ServerRequestExecutor serverRequestExecutor, Map<String, String> context) {
 		logger.debug(String.format(REGISTERED, serverRequestExecutor.getClass().getName()));
-		String requestExecutorModeId = context.get(RequestParameters.MODE);
+		String requestExecutorModeId = context.get(LicensingRequests.MODE);
 		if (requestExecutorModeId != null) {
 			serverRequestExecutor.setAccessModeId(requestExecutorModeId);
 			if (!requestExecutors.containsKey(requestExecutorModeId)) {
@@ -106,13 +106,13 @@ public class ServerRunnerImpl implements ServerRunner {
 			}
 		} else {
 			logger.error(REQUEST_HANDLER_NOT_FOUND + serverRequestExecutor.toString()
-					+ RequestParameters.MODE + requestExecutorModeId);
+					+ LicensingRequests.MODE + requestExecutorModeId);
 			;
 		}
 	}
 
 	public void unbindServerRequestExecutor(ServerRequestExecutor serverRequestExecutor, Map<String, String> context) {
-		String requestExecutorModeId = context.get(RequestParameters.MODE);
+		String requestExecutorModeId = context.get(LicensingRequests.MODE);
 		if (requestExecutorModeId != null) {
 			requestExecutors.remove(requestExecutorModeId, serverRequestExecutor);
 			for (Entry<String, ServerRequestHandler> entry : requestHandlers.entrySet()) {
@@ -143,7 +143,7 @@ public class ServerRunnerImpl implements ServerRunner {
 	public void bindServerRequestActions(ServerRequestAction action, Map<String, String> context) {
 		logger.debug(String.format(REGISTERED, action.getClass().getName()));
 
-		String actionId = context.get(RequestParameters.ACTION);
+		String actionId = context.get(LicensingRequests.ACTION);
 		if (actionId != null) {
 			if (!this.requestActions.containsKey(actionId)) {
 				this.requestActions.put(actionId, action);
@@ -155,7 +155,7 @@ public class ServerRunnerImpl implements ServerRunner {
 	public void unbindServerRequestActions(ServerRequestAction action, Map<String, String> context) {
 		logger.debug(action.getClass().getName());
 
-		String actionId = context.get(RequestParameters.ACTION);
+		String actionId = context.get(LicensingRequests.ACTION);
 		if (actionId != null) {
 			if (this.requestActions.containsKey(actionId)) {
 				this.requestActions.remove(actionId, action);
